[Feat] 필터 api연결  #1254
